refactor(cart): move cart order reducer
