+++ libs/cart/src/reducers/cart-order/cart-order-state.interface.ts
@@ -1,4 +1,4 @@
-import { DaffCartOrderResult } from '../models/cart-order-result';
+import { DaffCartOrderResult } from '../../models/cart-order-result';
 
 export interface DaffCartOrderReducerState<T extends DaffCartOrderResult = DaffCartOrderResult> {
   cartOrderResult: T;
