+++ libs/cart/src/reducers/cart-reducers.ts
@@ -1,5 +1,5 @@
 import { daffCartReducer } from './cart.reducer';
-import { daffCartOrderReducer } from './cart-order.reducer';
+import { daffCartOrderReducer } from './cart-order/cart-order.reducer';
 
 export const daffCartReducers = {
   cart: daffCartReducer,
